refact job_metadata api
