+++ src/dirbs/api/v1/__init__.py
@@ -35,18 +35,18 @@ from flask_apispec import use_kwargs, marshal_with, doc
 from werkzeug.exceptions import BadRequest
 from marshmallow import fields, validate
 
-from dirbs.api.common import job_metadata as job_metadata_common
 from dirbs.api.v1.resources import imei as imei_resource
 from dirbs.api.v1.resources import version as version_resource
 from dirbs.api.v1.resources import tac as tac_resource
 from dirbs.api.v1.resources import msisdn as msisdn_resource
 from dirbs.api.v1.resources import catalog as catalog_resource
+from dirbs.api.v1.resources import job_metadata as job_resource
+from dirbs.api.v1.schemas.job_metadata import JobMetadata, JobMetadataArgs
 from dirbs.api.v1.schemas.msisdn import MSISDN
 from dirbs.api.v1.schemas.tac import GSMATacInfo
 from dirbs.api.v1.schemas.catalog import Catalog, CatalogArgs
 from dirbs.api.v1.schemas.version import Version
 from dirbs.api.v1.schemas.imei import IMEI, IMEIArgs
-from dirbs.api.common.job_metadata import JobMetadata, JobMetadataArgs
 
 
 api = Blueprint('v1', __name__.split('.')[0])
@@ -147,7 +147,7 @@ def catalog_api(**kwargs):
 @disable_options_method()
 def job_metadata_api(**kwargs):
     """Job Metadata API route."""
-    return job_metadata_common.api(**kwargs)
+    return job_resource.job_metadata_api(**kwargs)
 
 
 @doc(description='Information about the code and DB schema version used by Core and presence of '
